W7. Point

+++ src/com/company/w7/Point.java
@@ -1,32 +1,32 @@
 package com.company.w7;
 
 public class Point {
-    private float x;
-    private float y;
+    private int x;
+    private int y;
 
-    public Point(){
-        setX(0.0f);
-        setY(0.0f);
+    public Point() {
+        setX(0);
+        setY(0);
     }
 
-    public Point(float x, float y){
+    public Point(int x, int y) {
         setX(x);
         setY(y);
     }
 
-    public float getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public float getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(int y) {
         this.y = y;
     }
 
@@ -34,14 +34,13 @@ public class Point {
         return new Point(x, y);
     }
 
-    public void setXY(float x,float y){
+    public void setXY(int x, int y) {
         setX(x);
         setY(y);
     }
 
-
     @Override
     public String toString() {
-        return "(" + x +","+ y + ")";
+        return "(" + x + "," + y + ")";
     }
 }
